[data] PR 2: Refactor IcebergDatasink onto TableDatasink + IcebergAdapter
Migrates the existing Iceberg write path onto the table abstraction
introduced in PR 1. No user-facing API changes; Dataset.write_iceberg and
the IcebergDatasink class import path are preserved.

* New: python/ray/data/_internal/datasource/iceberg_adapter.py exposes
  IcebergAdapter, a TableAdapter[DataFile] that owns the Iceberg-
  specific responsibilities -- catalog load, schema evolution via
  union_by_name, per-block Parquet writes via PyIceberg's
  _dataframe_to_data_files, copy-on-write UPSERT via create_match_filter,
  and transactional commit through PyIceberg's Table.transaction.

* Rewritten: iceberg_datasink.py. IcebergDatasink keeps its public
  constructor signature but is now a thin subclass of TableDatasink
  that builds an IcebergAdapter and forwards the lifecycle to the
  framework. Backwards-compat properties (table_identifier, _table,
  _table_metadata) are surfaced so existing tests that introspect the
  instance keep working.

* Removed: IcebergWriteResult dataclass (replaced by
  TableWriteTaskResult).

Iceberg keeps using `pyiceberg._dataframe_to_data_files` for the actual
file writes -- migrating it to ParquetFileWriter is a future PR. The
existing test_iceberg.py suite must pass unchanged; that's the proof
that the abstraction is behaviour-preserving against a mature format.
